+++ core/src/main/java/org/testcontainers/containers/DockerComposeContainer.java
@@ -395,6 +395,10 @@ interface DockerCompose {
  * Use Docker Compose container.
  */
 class ContainerisedDockerCompose extends GenericContainer<ContainerisedDockerCompose> implements DockerCompose {
+
+    private static final String DOCKER_SOCKET_PATH = "/var/run/docker.sock";
+    public static final char UNIX_PATH_SEPERATOR = ':';
+
     public ContainerisedDockerCompose(List<File> composeFiles, String identifier) {
 
         super(TestcontainersConfiguration.getInstance().getDockerComposeContainerImage());
@@ -405,14 +409,14 @@ class ContainerisedDockerCompose extends GenericContainer<ContainerisedDockerCom
         // Map the docker compose file into the container
         final File dockerComposeBaseFile = composeFiles.get(0);
         final String pwd = dockerComposeBaseFile.getAbsoluteFile().getParentFile().getAbsolutePath();
-        final String containerPwd = MountableFile.forHostPath(pwd).getResolvedPath();
+        final String containerPwd = MountableFile.forHostPath(pwd).getFilesystemPath();
 
         final List<String> absoluteDockerComposeFiles = composeFiles.stream()
                         .map(File::getAbsolutePath)
                         .map(MountableFile::forHostPath)
-                        .map(MountableFile::getResolvedPath)
+                        .map(MountableFile::getFilesystemPath)
                         .collect(toList());
-        final String composeFileEnvVariableValue = Joiner.on(File.pathSeparator).join(absoluteDockerComposeFiles);
+        final String composeFileEnvVariableValue = Joiner.on(UNIX_PATH_SEPERATOR).join(absoluteDockerComposeFiles); // we always need the UNIX path separator
         logger().debug("Set env COMPOSE_FILE={}", composeFileEnvVariableValue);
         addEnv(ENV_COMPOSE_FILE, composeFileEnvVariableValue);
         addFileSystemBind(pwd, containerPwd, READ_ONLY);
@@ -421,10 +425,16 @@ class ContainerisedDockerCompose extends GenericContainer<ContainerisedDockerCom
         //  as the docker daemon, just mapping the docker control socket is OK.
         // As there seems to be a problem with mapping to the /var/run directory in certain environments (e.g. CircleCI)
         //  we map the socket file outside of /var/run, as just /docker.sock
-        addFileSystemBind("/var/run/docker.sock", "/docker.sock", READ_WRITE);
+        addFileSystemBind(getDockerSocketHostPath(), "/docker.sock", READ_WRITE);
         addEnv("DOCKER_HOST", "unix:///docker.sock");
         setStartupCheckStrategy(new IndefiniteWaitOneShotStartupCheckStrategy());
         setWorkingDirectory(containerPwd);
+    }
+
+    private String getDockerSocketHostPath() {
+        return SystemUtils.IS_OS_WINDOWS
+                ? "/" + DOCKER_SOCKET_PATH
+                : DOCKER_SOCKET_PATH;
     }
 
     @Override